[Intermediate][Refactoring] Extracted start event and activity compilers
... from monolith compilerApp and gave the common superclass
